+++ src/modules/edc-demo/components/error-state/error-state.component.ts
@@ -13,5 +13,8 @@ export class ErrorStateComponent {
   cls = true;
 
   @Input()
+  error: any;
+
+  @Input()
   errorMessage = '';
 }